+++ packages/shiki/src/highlighter.ts
@@ -154,3 +154,18 @@ export interface Highlighter {
 
   // codeToImage?(): string
 }
+
+function repairThemeIfNoFallbackColor(theme: IShikiTheme) {
+  // Has the default no-scope setting with fallback colors
+  if (theme.settings[0].settings && !theme.settings[0].scope) {
+    return
+  }
+
+  // Push a no-scope setting with fallback colors
+  theme.settings.unshift({
+    settings: {
+      foreground: theme.fg,
+      background: theme.bg
+    }
+  })
+}