+++ scripts/pullGHThemes.sh
@@ -1,6 +1,0 @@
-rm -rf tmp
-mkdir -p tmp/themes
-node scripts/pullThemesFromGitHub.js
-node scripts/normalizeThemePaths.js
-cp tmp/themes/*.json packages/themes/data
-node scripts/updateThemeSrc.js